+1 New File/Indentation
Created a new html file called Photo gallery page.html which will be the page that opens up when the view more link for the photography section is clicked, and will show an all of the photos instead of just four. I also Properly indented the Html files for both index.html and Photo gallery page.html.